Move Reaction config to separate file
