+++ TwitchChannelPointsMiner/constants.py
@@ -1,7 +1,4 @@
 # Twitch endpoints
-from enum import Enum
-
-
 URL = "https://www.twitch.tv"
 IRC = "irc.chat.twitch.tv"
 IRC_PORT = 6667
@@ -210,10 +207,3 @@ class GQLOperations:
             }
         },
     }
-
-class Reaction(Enum):
-    HYPE = "HYPE",
-    FUNNY = "FUNNY",
-    LOVE = "LOVE",
-    WHAAAT = "WHAAAT",
-    OH_NO = "OH_NO",
